Open eyetracker datafile with session id
This makes it possible to have the session id in the name of the edf
file on the host of the eyelink pc. This makes it harder to directly
overwrite the name on the host. Allowing a longer grace period of
retrieving the file from the host, before its overwritten.

fixes #5